change scripts' numbers and import configuration
