+++ Quantum-Control-Applications/Superconducting/Two-Fixed-Coupled-Transmons/04_resonator_spectroscopy_multiplexed.py
@@ -17,7 +17,7 @@ Before proceeding to the next node:
 
 from qm.qua import *
 from qm import QuantumMachinesManager, SimulationConfig
-from configuration_mw_fem import *
+from configuration import *
 from qualang_tools.results import progress_counter, fetching_tool
 from qualang_tools.plot import interrupt_on_close
 from qualang_tools.loops import from_array
