+++ Quantum-Control-Applications/Superconducting/Two-Fixed-Coupled-Transmons/10a_readout_optimization_freq.py
@@ -18,7 +18,7 @@ Next steps before going to the next node:
 from qm import QuantumMachinesManager
 from qm.qua import *
 from qm import SimulationConfig
-from configuration_mw_fem import *
+from configuration import *
 import matplotlib.pyplot as plt
 from qualang_tools.loops import from_array
 from qualang_tools.results import fetching_tool, progress_counter
