+++ Quantum-Control-Applications/Superconducting/Two-Fixed-Coupled-Transmons/10b_readout_optimization_amp.py
@@ -17,7 +17,7 @@ Next steps before going to the next node:
 
 from qm import QuantumMachinesManager, SimulationConfig
 from qm.qua import *
-from configuration_mw_fem import *
+from configuration import *
 import matplotlib.pyplot as plt
 from qualang_tools.loops import from_array
 from qualang_tools.results import fetching_tool, progress_counter
