+++ Quantum-Control-Applications/Superconducting/Two-Fixed-Coupled-Transmons/14_T2ramsey.py
@@ -16,35 +16,35 @@ Next steps before going to the next node:
 
 from qm import QuantumMachinesManager, SimulationConfig
 from qm.qua import *
-from configuration_mw_fem import *
+from configuration import *
 import matplotlib.pyplot as plt
 from qualang_tools.loops import from_array
 from qualang_tools.results import fetching_tool
 from qualang_tools.plot import interrupt_on_close
 from qualang_tools.results import progress_counter
-from macros import qua_declaration, multiplexed_readout
+from macros import qua_declaration, multiplexed_readout, active_reset
+import math
 from qualang_tools.results.data_handler import DataHandler
+import matplotlib
 
-##################
-#   Parameters   #
-##################
-# Parameters Definition
-n_avg = 10  # The number of averages
-# Qubit detuning sweep with respect to qubit_IF
-freq_span = 10e6
-freq_step = 0.2e6
-dfs = np.arange(-freq_span, +freq_span, freq_step)
+matplotlib.use("TkAgg")
 
-t_max = 200
+###################
+# The QUA program #
+##################
+n_avg = 8 * 60  # The number of averages
+t_max = 8_000
 t_min = 4
-t_step = 6
-t_delay = np.arange(t_min, t_max, t_step)  # Idle time sweep in clock cycles (Needs to be a list of integers)
+t_step = 120
+t_delays = np.arange(t_min, t_max, t_step)  # Idle time sweep in clock cycles (Needs to be a list of integers)
+freq_detuning = 0.5 * u.MHz
+delta_phase = 4e-09 * freq_detuning * t_step
 
 # Data to save
 save_data_dict = {
     "n_avg": n_avg,
-    "dfs": dfs,
-    "t_delay": t_delay,
+    "t_delays": t_delays,
+    "detuning": freq_detuning,
     "config": config,
 }
 
@@ -54,43 +54,45 @@ save_data_dict = {
 
 with program() as PROGRAM:
     I, I_st, Q, Q_st, n, n_st = qua_declaration(nb_of_qubits=2)
+    state = [declare(bool) for _ in range(2)]
     t = declare(int)  # QUA variable for the idle time
     df = declare(int)  # QUA variable for the qubit frequency
+    phase = declare(fixed)
 
     with for_(n, 0, n < n_avg, n + 1):
-        with for_(*from_array(df, dfs)):
-            # Update the frequency of the two qubit elements
-            update_frequency("q1_xy", df + qubit_IF_q1)
-            update_frequency("q2_xy", df + qubit_IF_q2)
-
-            with for_(*from_array(t, t_delay)):
-                # qubit 1
-                play("x90", "q1_xy")
-                wait(t, "q1_xy")
-                play("x90", "q1_xy")
-
-                # qubit 2
-                play("x90", "q2_xy")
-                wait(t, "q2_xy")
-                play("x90", "q2_xy")
-
-                # Align the elements to measure after having waited a time "tau" after the qubit pulses.
-                align()
-                # Measure the state of the resonators
-                multiplexed_readout(I, I_st, Q, Q_st, resonators=[1, 2], weights="rotated_")
-                # Wait for the qubit to decay to the ground state
-                wait(thermalization_time * u.ns)
         # Save the averaging iteration to get the progress bar
         save(n, n_st)
+        assign(phase, 0)
+
+        with for_(*from_array(t, t_delays)):
+
+            assign(phase, phase + delta_phase)
+
+            play("x90", "q1_xy")
+            play("x90", "q2_xy")
+            wait(t)
+            frame_rotation_2pi(phase, "q1_xy")
+            frame_rotation_2pi(phase, "q2_xy")
+            play("x90", "q1_xy")
+            play("x90", "q2_xy")
+
+            # Align the elements to measure after having waited a time "tau" after the qubit pulses.
+            align()
+
+            # Measure the state of the resonators
+            multiplexed_readout(I, I_st, Q, Q_st, resonators=[1, 2], weights="rotated_")
+
+            wait(thermalization_time * u.ns)
+
+            reset_frame("q1_xy")
+            reset_frame("q2_xy")
 
     with stream_processing():
-        n_st.save("n")
-        # resonator 1
-        I_st[0].buffer(len(t_delay)).buffer(len(dfs)).average().save("I1")
-        Q_st[0].buffer(len(t_delay)).buffer(len(dfs)).average().save("Q1")
-        # resonator 2
-        I_st[1].buffer(len(t_delay)).buffer(len(dfs)).average().save("I2")
-        Q_st[1].buffer(len(t_delay)).buffer(len(dfs)).average().save("Q2")
+        n_st.save("iteration")
+        for ind in range(2):
+            I_st[ind].buffer(len(t_delays)).average().save(f"I{ind+1}")
+            Q_st[ind].buffer(len(t_delays)).average().save(f"Q{ind+1}")
+
 
 #####################################
 #  Open Communication with the QOP  #
@@ -119,7 +121,7 @@ else:
         fig = plt.figure()
         interrupt_on_close(fig, job)
         # Tool to easily fetch results from the OPX (results_handle used in it)
-        results = fetching_tool(job, ["n", "I1", "Q1", "I2", "Q2"], mode="live")
+        results = fetching_tool(job, ["iteration", "I1", "Q1", "I2", "Q2"], mode="live")
         # Live plotting
         while results.is_processing():
             # Fetch results
@@ -130,27 +132,25 @@ else:
             # Progress bar
             progress_counter(n, n_avg, start_time=results.start_time)
             # Plot
-            plt.suptitle("Ramsey chevron")
             plt.subplot(221)
             plt.cla()
-            plt.pcolor(4 * t_delay, dfs / u.MHz, I1)
-            plt.title(f"qubit 1 I, fcent={(qubit_LO_q1 + qubit_IF_q1) / u.MHz} MHz")
-            plt.ylabel("Frequency detuning [MHz]")
+            plt.plot(4 * t_delays, I1)
+            plt.ylabel("I quadrature [V]")
+            plt.title("Qubit 1")
             plt.subplot(223)
             plt.cla()
-            plt.pcolor(4 * t_delay, dfs / u.MHz, Q1)
-            plt.title("qubit 1 Q")
-            plt.xlabel("Idle time [ns]")
-            plt.ylabel("Frequency detuning [MHz]")
+            plt.plot(4 * t_delays, Q1)
+            plt.ylabel("Q quadrature [V]")
+            plt.xlabel("Idle times [ns]")
             plt.subplot(222)
             plt.cla()
-            plt.pcolor(4 * t_delay, dfs / u.MHz, I2)
-            plt.title(f"qubit 2 I, fcent={(qubit_LO_q2 + qubit_IF_q2) / u.MHz} MHz")
+            plt.plot(4 * t_delays, I2)
+            plt.title("Qubit 2")
             plt.subplot(224)
             plt.cla()
-            plt.pcolor(4 * t_delay, dfs / u.MHz, Q2)
-            plt.title("qubit 2 Q")
-            plt.xlabel("Idle time [ns]")
+            plt.plot(4 * t_delays, Q2)
+            plt.title("Q2")
+            plt.xlabel("Idle times [ns]")
             plt.tight_layout()
             plt.pause(1)
 
@@ -159,7 +159,7 @@ else:
         data_handler = DataHandler(root_data_folder=save_dir)
         save_data_dict.update({"fig_live": fig})
         data_handler.additional_files = {script_name: script_name, **default_additional_files}
-        data_handler.save_data(data=save_data_dict, name="ramsey_chevron")
+        data_handler.save_data(data=save_data_dict, name="ramsey")
 
     except Exception as e:
         print(f"An exception occurred: {e}")
