+++ Quantum-Control-Applications/Superconducting/Two-Fixed-Coupled-Transmons/17a_single_qubit_RB.py
@@ -1,11 +1,11 @@
 """
-        SINGLE QUBIT INTERLEAVED RANDOMIZED BENCHMARKING
+        SINGLE QUBIT RANDOMIZED BENCHMARKING
 """
 
 from qm.qua import *
 from qm import QuantumMachinesManager
 from qm import SimulationConfig
-from configuration_mw_fem import *
+from configuration import *
 import matplotlib.pyplot as plt
 import numpy as np
 from qualang_tools.bakery.randomized_benchmark_c1 import c1_table
@@ -14,23 +14,23 @@ from qualang_tools.plot import interrupt_on_close
 from macros import multiplexed_readout, qua_declaration, active_reset
 import math
 from qualang_tools.results.data_handler import DataHandler
+from scipy.optimize import curve_fit
+import matplotlib
+import time
+
+matplotlib.use("TkAgg")
 
 ##################
 #   Parameters   #
 ##################
 # Parameters Definition
-num_of_sequences = 10  # Number of random sequences
-n_avg = 3  # Number of averaging loops for each random sequence
-max_circuit_depth = 20  # Maximum circuit depth
-delta_clifford = 2  #  Play each sequence with a depth step equals to 'delta_clifford - Must be > 1
+num_of_sequences = 50  # Number of random sequences
+n_avg = 10  # Number of averaging loops for each random sequence
+max_circuit_depth = 200  # Maximum circuit depth
+delta_clifford = 20  #  Play each sequence with a depth step equals to 'delta_clifford - Must be > 1
 seed = 345324  # Pseudo-random number generator seed
 # List of recovery gates from the lookup table
 inv_gates = [int(np.where(c1_table[i, :] == 0)[0][0]) for i in range(24)]
-# index of the gate to interleave from the play_sequence() function defined below
-# Correspondence table:
-#  0: identity |  1: x180 |  2: y180
-# 12: x90      | 13: -x90 | 14: y90 | 15: -y90 |
-interleaved_gate_index = 0
 
 # Assertion
 assert (max_circuit_depth / delta_clifford).is_integer(), "max_circuit_depth / delta_clifford must be an integer."
@@ -42,51 +42,32 @@ save_data_dict = {
     "num_of_sequences": num_of_sequences,
     "delta_clifford": delta_clifford,
     "max_circuit_depth": max_circuit_depth,
-    "interleaved_gate_index": interleaved_gate_index,
 }
 
 
 ###################################
 # Helper functions and QUA macros #
 ###################################
-def get_interleaved_gate(gate_index):
-    if gate_index == 0:
-        return "I"
-    elif gate_index == 1:
-        return "x180"
-    elif gate_index == 2:
-        return "y180"
-    elif gate_index == 12:
-        return "x90"
-    elif gate_index == 13:
-        return "-x90"
-    elif gate_index == 14:
-        return "y90"
-    elif gate_index == 15:
-        return "-y90"
+def power_law(power, a, b, p):
+    return a * (p**power) + b
 
 
-def generate_sequence(interleaved_gate_index):
+def generate_sequence():
     cayley = declare(int, value=c1_table.flatten().tolist())
     inv_list = declare(int, value=inv_gates)
     current_state = declare(int)
     step = declare(int)
-    sequence = declare(int, size=2 * max_circuit_depth + 1)
-    inv_gate = declare(int, size=2 * max_circuit_depth + 1)
+    sequence = declare(int, size=max_circuit_depth + 1)
+    inv_gate = declare(int, size=max_circuit_depth + 1)
     i = declare(int)
     rand = Random(seed=seed)
 
     assign(current_state, 0)
-    with for_(i, 0, i < 2 * max_circuit_depth, i + 2):
+    with for_(i, 0, i < max_circuit_depth, i + 1):
         assign(step, rand.rand_int(24))
         assign(current_state, cayley[current_state * 24 + step])
         assign(sequence[i], step)
         assign(inv_gate[i], inv_list[current_state])
-        # interleaved gate
-        assign(step, interleaved_gate_index)
-        assign(current_state, cayley[current_state * 24 + step])
-        assign(sequence[i + 1], step)
-        assign(inv_gate[i + 1], inv_list[current_state])
 
     return sequence, inv_gate
 
@@ -96,7 +77,7 @@ def play_sequence(sequence_list, depth, qb):
     with for_(i, 0, i <= depth, i + 1):
         with switch_(sequence_list[i], unsafe=True):
             with case_(0):
-                wait(pi_len >> 2, qb)
+                wait(pi_len // 4, qb)
             with case_(1):
                 play("x180", qb)
             with case_(2):
@@ -184,34 +165,34 @@ with program() as PROGRAM:
     with for_(m, 0, m < num_of_sequences, m + 1):  # QUA for_ loop over the random sequences
         # Save the counter for the progress bar
         save(m, m_st)
-        # Generates the RB sequence with a gate interleaved after each Clifford
-        sequence_list, inv_gate_list = generate_sequence(interleaved_gate_index=interleaved_gate_index)
-        # Depth_target is used to always play the gates by pairs [(random_gate-interleaved_gate)^depth/2-inv_gate]
-        assign(depth_target, 0)
+        sequence_list, inv_gate_list = generate_sequence()  # Generate the random sequence of length max_circuit_depth
 
-        with for_(depth, 1, depth <= 2 * max_circuit_depth, depth + 1):
+        assign(depth_target, 0)  # Initialize the current depth to 0
+
+        with for_(depth, 1, depth <= max_circuit_depth, depth + 1):  # Loop over the depths
             # Replacing the last gate in the sequence with the sequence's inverse gate
             # The original gate is saved in 'saved_gate' and is being restored at the end
             assign(saved_gate, sequence_list[depth])
             assign(sequence_list[depth], inv_gate_list[depth - 1])
             # Only played the depth corresponding to target_depth
 
-            with if_((depth == 2) | (depth == depth_target)):
+            with if_((depth == 1) | (depth == depth_target)):
 
                 with for_(n, 0, n < n_avg, n + 1):  # Averaging loop
-                    # Can replace by active reset
                     wait(thermalization_time * u.ns)
                     # Align the two elements to play the sequence after qubit initialization
                     align()
                     # The strict_timing ensures that the sequence will be played without gaps
+                    # Play the random sequence of desired depth
                     play_sequence(sequence_list, depth, "q1_xy")
                     play_sequence(sequence_list, depth, "q2_xy")
                     # Align the two elements to measure after playing the circuit.
                     align()
                     multiplexed_readout(I, I_st, Q, Q_st, resonators=[1, 2], weights="rotated_")
-                # always play the random gate followed by the interleaved gate. The factor of 2 is there to always
-                # play the gates by pairs [(random_gate-interleaved_gate)^depth/2-inv_gate]
-                assign(depth_target, depth_target + 2 * delta_clifford)
+
+                # Go to the next depth
+                assign(depth_target, depth_target + delta_clifford)
+
             # Reset the last gate of the sequence back to the original Clifford gate
             # (that was replaced by the recovery gate at the beginning)
             assign(sequence_list[depth], saved_gate)
@@ -244,7 +225,8 @@ if simulate:
     job.get_simulated_samples().con1.plot()
     plt.plot(block=False)
 else:
-    try:  # Open the quantum machine
+    try:
+        # Open the quantum machine
         qm = qmm.open_qm(config)
 
         # Send the QUA program to the OPX, which compiles and executes it
@@ -257,7 +239,7 @@ else:
         results = fetching_tool(job, ["iteration", "I1", "Q1", "I2", "Q2"], mode="live")
         # Live plotting
 
-        x = np.arange(0, 2 * max_circuit_depth + 0.1, 2 * delta_clifford)
+        x = np.arange(0, max_circuit_depth + 0.1, delta_clifford)
         x[0] = 1  # to set the first value of 'x' to be depth = 1 as in the experiment
 
         while results.is_processing():
@@ -267,7 +249,8 @@ else:
             # Progress bar
             progress_counter(res[0], num_of_sequences, start_time=results.start_time)
 
-            plt.suptitle("interleaved-RB")
+            plt.suptitle("RB")
+
             for ind in range(2):
 
                 S = res[2 * ind + 1]
@@ -282,16 +265,12 @@ else:
             plt.tight_layout()
             plt.pause(2)
 
-        for ind in range(2):
-            save_data_dict[f"I{ind}"] = res[2 * ind + 1]
-            save_data_dict[f"Q{ind}"] = res[2 * ind + 2]
-
         # Save results
         script_name = Path(__file__).name
         data_handler = DataHandler(root_data_folder=save_dir)
         save_data_dict.update({"fig_live": fig})
         data_handler.additional_files = {script_name: script_name, **default_additional_files}
-        data_handler.save_data(data=save_data_dict, name="rb_interleaved")
+        data_handler.save_data(data=save_data_dict, name="rb")
 
     except Exception as e:
         print(f"An exception occurred: {e}")
