+++ Quantum-Control-Applications/Superconducting/Two-Fixed-Coupled-Transmons/19b_CR_calib_cr_drive_amplitude.py
@@ -39,7 +39,7 @@ Reference: Sarah Sheldon, Easwar Magesan, Jerry M. Chow, and Jay M. Gambetta Phy
 
 from qm.qua import *
 from qm import QuantumMachinesManager
-from configuration_mw_fem import *
+from configuration import *
 import matplotlib.pyplot as plt
 from qm import SimulationConfig
 from qualang_tools.loops import from_array
