+++ Quantum-Control-Applications/Superconducting/Two-Fixed-Coupled-Transmons/19d_CR_calib_cr_cancel_phase.py
@@ -1,5 +1,5 @@
 """
-                                 CR_calib_cancel_drive_amplitude
+                                 CR_calib_cancel_drive_phase
 
 The CR_calib scripts are designed for calibrating cross-resonance (CR) gates involving a system
 with a control qubit and a target qubit. These scripts help estimate the parameters of a Hamiltonian,
@@ -32,24 +32,24 @@ Prerequisites:
     - (optional) Having calibrated the readout (readout_frequency, amplitude, duration_optimization IQ_blobs) for better SNR.
 
 Next steps before going to the next node:
-    - Find the amplitude where a_X (coeff of I_X) and a_Y (coeff of I_Y) is zero simultaneously.
-      If the two coeffs do not vanish simultaneously, the value of phi could be wrong.
-      Set cr_c1t2_drive_am in the configuration file.
-      Set cr_cancel_c1t2_square_positive_amp in the configuration file accordingly.
+    - Find the phase to shift for the CR cancel drive via phi = arctan(IY/IX).
+      Alternatively, find the phase where a_Y (coeff of I_Y) is zero. We call it phi1.
+      phi = phi0 - phi1.
+      Note that the phase is in units of 2 * pi as it is used with `frame_rotation_2pi`.
 
 Reference: Sarah Sheldon, Easwar Magesan, Jerry M. Chow, and Jay M. Gambetta Phys. Rev. A 93, 060302(R) (2016)
 """
 
 from qm.qua import *
 from qm import QuantumMachinesManager
-from configuration_mw_fem import *
+from configuration import *
 import matplotlib.pyplot as plt
 from qm import SimulationConfig
 from qualang_tools.loops import from_array
 from qualang_tools.results import fetching_tool
 from qualang_tools.plot import interrupt_on_close
 from qualang_tools.results import progress_counter
-from macros import qua_declaration, multiplexed_readout, active_reset
+from macros import qua_declaration, multiplexed_readout
 from qualang_tools.results.data_handler import DataHandler
 from macros import qua_declaration, multiplexed_readout
 from cr_hamiltonian_tomography import (
@@ -76,7 +76,10 @@ cr_drive_phase = 0.0  # in units of 2pi
 cr_cancel_amp = 0.5  # ratio
 cr_cancel_phase = 0.0  # in units of 2pi
 ts_cycles = np.arange(4, 100, 1)  # in clock cylcle = 4ns
-amp_scalings = np.arange(0.5, 1.01, 0.05)  # scaling factor for amplitude
+phases = np.arange(0.0, 1.01, 0.05)  # ratio relative to 2 * pi
+
+# Readout Parameters
+weights = "rotated_"  # ["", "rotated_", "opt_"]
 
 # Derived parameters
 qc_xy = f"q{qc}_xy"
@@ -100,7 +103,7 @@ save_data_dict = {
     "cr_cancel_amp": cr_cancel_amp,
     "cr_cancel_phase": cr_cancel_phase,
     "ts_ns": ts_ns,
-    "amp_scalings": amp_scalings,
+    "phases": phases,
     "n_avg": n_avg,
     "config": config,
 }
@@ -115,16 +118,13 @@ with program() as PROGRAM:
     state = [declare(bool) for _ in range(2)]
     state_st = [declare_stream() for _ in range(2)]
     t = declare(int)
-    a = declare(fixed)
-    counter = declare(int, value=0)
-    counter_st = declare_stream()
+    ph = declare(fixed)
     s = declare(int)  # QUA variable for the control state
     c = declare(int)  # QUA variable for the projection index in QST
 
-    with for_(*from_array(a, amp_scalings)):
-        save(counter, counter_st)
-        wait(60 * u.ms)
-        with for_(n, 0, n < n_avg, n + 1):
+    with for_(n, 0, n < n_avg, n + 1):
+        save(n, n_st)
+        with for_(*from_array(ph, phases)):
             with for_(*from_array(t, ts_cycles)):
                 with for_(c, 0, c < 3, c + 1):  # bases
                     with for_(s, 0, s < 2, s + 1):  # states
@@ -135,11 +135,11 @@ with program() as PROGRAM:
                         if cr_type == "direct+cancel":
                             # phase shift for cancel drive
                             frame_rotation_2pi(cr_drive_phase, cr_drive)
-                            frame_rotation_2pi(cr_cancel_phase, cr_cancel)
+                            frame_rotation_2pi(ph, cr_cancel)
                             # direct + cancel
                             align(qc_xy, cr_drive, cr_cancel)
-                            play("square_positive" * amp(cr_drive_amp), cr_drive, duration=t)
-                            play("square_positive" * amp(a), cr_cancel, duration=t)
+                            play("square_positive", cr_drive, duration=t)
+                            play("square_positive", cr_cancel, duration=t)
                             # align for the next step and clear the phase shift
                             align(qt_xy, cr_drive, cr_cancel)
                             reset_frame(cr_drive)
@@ -148,18 +148,18 @@ with program() as PROGRAM:
                         elif cr_type == "direct+cancel+echo":
                             # phase shift for cancel drive
                             frame_rotation_2pi(cr_drive_phase, cr_drive)
-                            frame_rotation_2pi(cr_cancel_phase, cr_cancel)
+                            frame_rotation_2pi(ph, cr_cancel)
                             # direct + cancel
                             align(qc_xy, cr_drive, cr_cancel)
-                            play("square_positive" * amp(cr_drive_amp), cr_drive, duration=t)
-                            play("square_positive" * amp(a), cr_cancel, duration=t)
+                            play("square_positive", cr_drive, duration=t)
+                            play("square_positive", cr_cancel, duration=t)
                             # pi pulse on control
                             align(qc_xy, cr_drive, cr_cancel)
                             play("x180", qc_xy)
                             # echoed direct + cancel
                             align(qc_xy, cr_drive, cr_cancel)
-                            play("square_negative" * amp(cr_drive_amp), cr_drive, duration=t)
-                            play("square_negative" * amp(a), cr_cancel, duration=t)
+                            play("square_negative", cr_drive, duration=t)
+                            play("square_negative", cr_cancel, duration=t)
                             # pi pulse on control
                             align(qc_xy, cr_drive, cr_cancel)
                             play("x180", qc_xy)
@@ -188,21 +188,21 @@ with program() as PROGRAM:
 
                         # Wait for the qubit to decay to the ground state - Can be replaced by active reset
                         wait(thermalization_time * u.ns)
-        assign(counter, counter + 1)
+
     with stream_processing():
-        counter_st.save("n")
+        n_st.save("n")
         # control qubit
-        I_st[0].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(amp_scalings)).average().save("I1")
-        Q_st[0].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(amp_scalings)).average().save("Q1")
-        state_st[0].boolean_to_int().buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(
-            len(amp_scalings)
-        ).average().save("state1")
+        I_st[0].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(phases)).average().save("I1")
+        Q_st[0].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(phases)).average().save("Q1")
+        state_st[0].boolean_to_int().buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(phases)).average().save(
+            "state1"
+        )
         # target qubit
-        I_st[1].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(amp_scalings)).average().save("I2")
-        Q_st[1].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(amp_scalings)).average().save("Q2")
-        state_st[1].boolean_to_int().buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(
-            len(amp_scalings)
-        ).average().save("state2")
+        I_st[1].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(phases)).average().save("I2")
+        Q_st[1].buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(phases)).average().save("Q2")
+        state_st[1].boolean_to_int().buffer(2).buffer(3).buffer(len(ts_cycles)).buffer(len(phases)).average().save(
+            "state2"
+        )
 
 
 #####################################
@@ -240,18 +240,17 @@ else:
             res = results.fetch_all()
             iterations, I1, Q1, state_c, I2, Q2, state_t = res
             # Progress bar
-            progress_counter(iterations, len(amp_scalings), start_time=results.start_time)
+            progress_counter(iterations, n_avg, start_time=results.start_time)
             # Convert the results into Volts
             I1, Q1 = u.demod2volts(I1, readout_len), u.demod2volts(Q1, readout_len)
             I2, Q2 = u.demod2volts(I2, readout_len), u.demod2volts(Q2, readout_len)
             bloch_c, bloch_t = -2 * state_c + 1, -2 * state_t + 1  # convert |0> -> 1, |1> -> -1
             # Progress bar
             progress_counter(iterations, n_avg, start_time=results.start_time)
-
-        # plotting data
-        fig = plot_cr_duration_vs_scan_param(bloch_c, bloch_t, ts_ns, amp_scalings, "cr cancel amplitude", axss)
-        plt.tight_layout()
-        # plt.pause(1)
+            # plotting data
+            fig = plot_cr_duration_vs_scan_param(bloch_c, bloch_t, ts_ns, phases, "cr cancel phase [2pi]", axss)
+            plt.tight_layout()
+            plt.pause(1)
 
         # Save data
         save_data_dict.update({"fig_live": fig})
@@ -261,35 +260,36 @@ else:
         # Perform CR Hamiltonian tomography
         coeffs = []
         fig_analyses = []
-        for idx, a in enumerate(amp_scalings):
+        for idx, ph in enumerate(phases):
             print("-" * 40)
-            print(f"fitting for amp = {a}")
+            print(f"fitting for phase = {ph}")
             try:
                 crht = CRHamiltonianTomographyAnalysis(
                     ts=ts_ns,
-                    data=bloch_t[idx, ...],  # target data: len(amp_scalings) x len(t_vec_cycle) x 3 x 2
+                    data=bloch_t[idx, ...],  # target data: len(phases) x len(t_vec_cycle) x 3 x 2
                 )
                 crht.fit_params()
                 coeffs.append(crht.interaction_coeffs_MHz)
                 fig_analysis = crht.plot_fit_result(do_show=False)
-                save_data_dict[f"fig_analysis_amp_scaling={a:5.4f}".replace(".", "-")] = fig_analysis
+                save_data_dict[f"fig_analysis_phase={ph:5.4f}".replace(".", "-")] = fig_analysis
             except:
-                print(f"-> failed")
+                print(f"fitting failed at phase = {ph}")
                 crht.interaction_coeffs_MHz = {k: None for k, v in crht.interaction_coeffs_MHz.items()}
                 coeffs.append({p: None for p in PAULI_2Q})
             finally:
-                save_data_dict[f"ham_tomo_params_fitted_@amp={a:5.4f}"] = crht.params_fitted_dict
-                save_data_dict[f"ham_tomo_interaction_coeffs_MHz_@amp={a:5.4f}"] = crht.interaction_coeffs_MHz
+                # plt.show(block=False)
+                # save_data_dict["fig_analysis"] = fig_analysis
+                save_data_dict[f"ham_tomo_params_fitted_@phase={ph:5.4f}"] = crht.params_fitted_dict
+                save_data_dict[f"ham_tomo_interaction_coeffs_MHz_@phase={ph:5.4f}"] = crht.interaction_coeffs_MHz
 
         # Plot the estimated interaction coefficients
-        fig_summary = plot_interaction_coeffs(coeffs, amp_scalings, xlabel="cr cancel amplitude")
+        fig_summary = plot_interaction_coeffs(coeffs, phases, xlabel="cr cancel phase")
         save_data_dict["fig_summary"] = fig_summary
 
         # Save results
         script_name = Path(__file__).name
         data_handler = DataHandler(root_data_folder=save_dir)
-        data_handler.additional_files = {script_name: script_name, **default_additional_files}
-        data_handler.save_data(data=save_data_dict, name="cr_calib_ham_tomo_cancel_vs_amp")
+        data_handler.save_data(data=save_data_dict, name="cr_calib_ham_tomo_cancel_vs_phase")
 
     except Exception as e:
         print(f"An exception occurred: {e}")
